fix: standardize utils, ErrorCode coverage, include order, DLL exports
- Add ErrorCode::PayloadTooLarge (413), UnsupportedMediaType (415),
  UnprocessableEntity (422), ValidationFailed (4221) + to_string coverage
- Fix kPbkdf2SaltBytes/HashBytes type: int → size_t (consistency)
- Fix ComputerRepository.cpp include order: project before third-party
- Fix validation_utils parameter naming: camelCase → snake_case
- Replace inline timestampToIso() with time_utils::format_iso8601()
- Wire sanitize_input() to use string_utils::trim() (no duplicate)
- Add HUB32API_EXPORT to all 4 utils modules for DLL visibility

+++ include/hub32api/core/Constants.hpp
@@ -81,8 +81,8 @@ constexpr size_t kSecretKeyBytes     = 32;   // bytes — JWT secret key length 
 constexpr size_t kUuidBytes          = 16;   // bytes — UUID v4 raw bytes
 constexpr size_t kTimestampBufSize   = 32;   // bytes — ISO 8601 timestamp buffer
 constexpr int    kPbkdf2Iterations   = 310000;  // rounds — OWASP 2024 minimum for PBKDF2-SHA256
-constexpr int    kPbkdf2SaltBytes    = 16;   // bytes — PBKDF2 salt length
-constexpr int    kPbkdf2HashBytes    = 32;   // bytes — PBKDF2 derived key length
+constexpr size_t kPbkdf2SaltBytes   = 16;    // bytes — PBKDF2 salt length
+constexpr size_t kPbkdf2HashBytes   = 32;    // bytes — PBKDF2 derived key length
 
 // -----------------------------------------------------------------------
 // ICMP
